data_encoder/requirements.txt
sentence-transformers
torch
git+https://github.com/openai/CLIP.git
pillow
bs4
tqdm
nltk
pysolr